refactor(InputView): 입력된 자동차의 이름을 검증하는 부분을 CarNameValidator 클래스로 분리
