+++ src/main/java/racingcar/StringUtil.java
@@ -1,0 +1,11 @@
+package racingcar;
+
+import static racingcar.view.InputView.*;
+
+public class StringUtil {
+	public static String[] splitUsingComma(String inputEveryCarsName) {
+		CarNameValidator.hasNoCarNameAtLast(inputEveryCarsName);
+		return inputEveryCarsName.split(CAR_SPLIT_REGEX);
+	}
+}
+
