+++ src/main/java/racingcar/view/InputView.java
@@ -1,7 +1,5 @@
 package racingcar.view;
 
-import java.util.Arrays;
-
 import camp.nextstep.edu.missionutils.Console;
 
 public class InputView {
@@ -18,35 +16,9 @@ public class InputView {
 	public static final String NUMBER_REGEX = "-?\\d+";
 	public static final int CAR_NAME_LETTERS_STANDARD = 5;
 
-	public static String[] enterEveryCarName() {
-		while (true) {
-			try {
-				System.out.println(INPUT_EVERY_CAR_NAME_MESSAGE);
-				String inputValue = Console.readLine();
-				hasNoCarNameAtLast(inputValue);
-				String[] everyCarName = inputValue.split(CAR_SPLIT_REGEX);
-				Arrays.stream(everyCarName).forEach(eachCarName -> validateCarName(eachCarName));
-				return everyCarName;
-			} catch (IllegalArgumentException e) {
-				System.out.println(String.format(SHOW_ERROR_MESSAGE, e.getMessage()));
-			}
-		}
-	}
-
-	private static void hasNoCarNameAtLast(String inputValue) {
-		if (inputValue.endsWith(CAR_SPLIT_REGEX)) {
-			throw new IllegalArgumentException(CAR_NAME_NOT_NULL_ERROR_MESSAGE);
-		}
-	}
-
-	private static void validateCarName(String eachCarName) {
-		if (eachCarName.length() > CAR_NAME_LETTERS_STANDARD) {
-			throw new IllegalArgumentException(
-				String.format(CAR_NAME_OVER_STANDARD_ERROR_MESSAGE, CAR_NAME_LETTERS_STANDARD));
-		}
-		if (eachCarName.isEmpty()) {
-			throw new IllegalArgumentException(CAR_NAME_NOT_NULL_ERROR_MESSAGE);
-		}
+	public static String enterEveryCarName() {
+		System.out.println(INPUT_EVERY_CAR_NAME_MESSAGE);
+		return Console.readLine();
 	}
 
 	public static int enterTryCnt() {
